Additional learned check for Sandsear Storm.
After the TM code is done, it runs the "does larvesta known the move" check AGAIN, because I forgot that if you select the new move in the TM menu, you can cancel it that way. Now it'll just immediately start the process over if you try to do that. This DID come at the cost of losing the "does Larvesta already know the move going in?" check... but we weren't actually doing anything about the info. Burrito was making noises about wanting to de-hardcode some of this... so if he wants me to add that check back in for his potential usage, I can rewrite this without too much difficulty.